requirements.txt
torch
torchvision
numpy
matplotlib
seaborn
scikit-learn
pandas
PyYAML
tqdm
tensorboard
pytest
flake8
